+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0 (unreleased)
 
+- Added `language` option
 - Dropped support for Ruby < 2.7
 
 ## 0.1.3 (2021-11-23)